+++ CHANGELOG.rst
@@ -21,6 +21,8 @@ not released yet
 * FIX `search` will no longer break on overwritten events with a master event
 * FIX when using short dates, khal infers that you meant next year, when date
   is before today
+* FIX Check for multi_uid .ics files in vdirs and don't import those events
+  (All .ics files in vdirs should only contain VEVENTS with the same UID.)
 
 * CHANGE only searched configuration file paths are now
   $XDG_CONFIG_HOME/khal/config and $XDG_CONFIG_HOME/khal/khal.conf (deprecated)